feat: add various aperture classes including Radial, Polygon, Elliptical, Rectangular, and OffsetRadial
