--- optiland/physical_apertures/physical_apertures.py
@@ -1,407 +1,0 @@
-"""Physical Apertures Module
-
-This module provides classes to define physical apertures on optical surfaces.
-For example, a telescope primary mirror with a center hole can be represented
-by a radial aperture that clips rays based on their distance from the origin.
-
-Kramer Harrison, 2024
-"""
-from abc import ABC, abstractmethod
-import numpy as np
-from matplotlib.path import Path
-
-
-class BaseAperture(ABC):
-    """
-    Base class for physical apertures.
-
-    Methods:
-        clip(RealRays): Clips the given rays based on the aperture's shape.
-    """
-    _registry = {}
-
-    def __init_subclass__(cls, **kwargs):
-        """Automatically register subclasses."""
-        super().__init_subclass__(**kwargs)
-        BaseAperture._registry[cls.__name__] = cls
-
-    def clip(self, rays):
-        """
-        Clips the given rays based on the aperture's shape.
-
-        Parameters:
-            rays (RealRays): List of rays to be clipped.
-
-        Returns:
-            list: List of clipped rays.
-        """
-        return rays  # pragma: no cover
-
-    @abstractmethod
-    def scale(self, scale_factor):
-        """
-        Scales the aperture by the given factor.
-
-        Parameters:
-            scale_factor (float): The factor by which to scale the aperture.
-        """
-        pass  # pragma: no cover
-
-    def to_dict(self):
-        """
-        Convert the aperture to a dictionary.
-
-        Returns:
-            dict: The dictionary representation of the aperture.
-        """
-        return {
-            'type': self.__class__.__name__
-        }
-
-    @classmethod
-    def from_dict(cls, data):
-        """
-        Create an aperture from a dictionary representation.
-
-        Parameters:
-            data (dict): The dictionary representation of the aperture.
-
-        Returns:
-            BaseAperture: The aperture object.
-        """
-        aperture_type = data['type']
-        return cls._registry[aperture_type].from_dict(data)
-
-
-class RadialAperture(BaseAperture):
-    """
-    Represents a radial aperture that clips rays based on their distance from
-    the origin.
-
-    Attributes:
-        r_max (float): The maximum radius allowed for the rays.
-        r_min (float): The minimum radius allowed for the rays. Defaults to 0.
-    """
-
-    def __init__(self, r_max, r_min=0):
-        super().__init__()
-        self.r_max = r_max
-        self.r_min = r_min
-
-    def clip(self, rays):
-        """
-        Clips the given rays based on their distance from the origin.
-
-        Args:
-            rays (Rays): The rays to be clipped.
-        """
-        radius2 = rays.x**2 + rays.y**2
-        condition = (radius2 > self.r_max**2) | (radius2 < self.r_min**2)
-        rays.clip(condition)
-
-    def scale(self, scale_factor):
-        """
-        Scales the aperture by the given factor.
-
-        Args:
-            scale_factor (float): The factor by which to scale the aperture.
-        """
-        self.r_max *= scale_factor
-        self.r_min *= scale_factor
-
-    def to_dict(self):
-        """
-        Convert the aperture to a dictionary.
-
-        Returns:
-            dict: The dictionary representation of the aperture.
-        """
-        aperture_dict = super().to_dict()
-        aperture_dict['r_max'] = self.r_max
-        aperture_dict['r_min'] = self.r_min
-        return aperture_dict
-
-    @classmethod
-    def from_dict(cls, data):
-        """
-        Create an aperture from a dictionary representation.
-
-        Args:
-            data (dict): The dictionary representation of the aperture.
-
-        Returns:
-            RadialAperture: The aperture object.
-        """
-        return cls(data['r_max'], data['r_min'])
-
-
-class OffsetRadialAperture(RadialAperture):
-    """
-    Represents a radial aperture that clips rays based on their distance from
-    the origin, with an offset in the x and y directions.
-
-    Attributes:
-        r_max (float): The maximum radius allowed for the rays.
-        r_min (float): The minimum radius allowed for the rays. Defaults to 0.
-        offset_x (float): The x-coordinate of the aperture's center.
-        offset_y (float): The y-coordinate of the aperture's center.
-    """
-
-    def __init__(self, r_max, r_min=0, offset_x=0, offset_y=0):
-        super().__init__(r_max, r_min)
-        self.offset_x = offset_x
-        self.offset_y = offset_y
-
-    def clip(self, rays):
-        """
-        Clips the given rays based on their distance from the origin.
-
-        Args:
-            rays (Rays): The rays to be clipped.
-        """
-        radius2 = (rays.x - self.offset_x)**2 + (rays.y - self.offset_y)**2
-        condition = (radius2 > self.r_max**2) | (radius2 < self.r_min**2)
-        rays.clip(condition)
-
-    def scale(self, scale_factor):
-        """
-        Scales the aperture by the given factor.
-
-        Args:
-            scale_factor (float): The factor by which to scale the aperture.
-        """
-        super().scale(scale_factor)
-        self.offset_x *= scale_factor
-        self.offset_y *= scale_factor
-
-    def to_dict(self):
-        """
-        Convert the aperture to a dictionary.
-
-        Returns:
-            dict: The dictionary representation of the aperture.
-        """
-        aperture_dict = super().to_dict()
-        aperture_dict['offset_x'] = self.offset_x
-        aperture_dict['offset_y'] = self.offset_y
-        return aperture_dict
-
-    @classmethod
-    def from_dict(cls, data):
-        """
-        Create an aperture from a dictionary representation.
-
-        Args:
-            data (dict): The dictionary representation of the aperture.
-
-        Returns:
-            OffsetRadialAperture: The aperture object.
-        """
-        return cls(data['r_max'], data['r_min'],
-                   data['offset_x'], data['offset_y'])
-
-
-class RectangularAperture(BaseAperture):
-    """
-    Represents a rectangular aperture that clips rays based on their position.
-
-    Attributes:
-        x_min (float): The minimum x-coordinate allowed for the rays.
-        x_max (float): The maximum x-coordinate allowed for the rays.
-        y_min (float): The minimum y-coordinate allowed for the rays.
-        y_max (float): The maximum y-coordinate allowed for the rays.
-    """
-
-    def __init__(self, x_min, x_max, y_min, y_max):
-        super().__init__()
-        self.x_min = x_min
-        self.x_max = x_max
-        self.y_min = y_min
-        self.y_max = y_max
-
-    def clip(self, rays):
-        """
-        Clips the given rays based on their position.
-
-        Args:
-            rays (Rays): The rays to be clipped.
-        """
-        condition = (rays.x < self.x_min) | (rays.x > self.x_max) | \
-                    (rays.y < self.y_min) | (rays.y > self.y_max)
-        rays.clip(condition)
-
-    def scale(self, scale_factor):
-        """
-        Scales the aperture by the given factor.
-
-        Args:
-            scale_factor (float): The factor by which to scale the aperture.
-        """
-        self.x_min *= scale_factor
-        self.x_max *= scale_factor
-        self.y_min *= scale_factor
-        self.y_max *= scale_factor
-
-    def to_dict(self):
-        """
-        Convert the aperture to a dictionary.
-
-        Returns:
-            dict: The dictionary representation of the aperture.
-        """
-        aperture_dict = super().to_dict()
-        aperture_dict['x_min'] = self.x_min
-        aperture_dict['x_max'] = self.x_max
-        aperture_dict['y_min'] = self.y_min
-        aperture_dict['y_max'] = self.y_max
-        return aperture_dict
-
-    @classmethod
-    def from_dict(cls, data):
-        """
-        Create an aperture from a dictionary representation.
-
-        Args:
-            data (dict): The dictionary representation of the aperture.
-
-        Returns:
-            RectangularAperture: The aperture object.
-        """
-        return cls(data['x_min'], data['x_max'],
-                   data['y_min'], data['y_max'])
-
-
-class EllipticalAperture(BaseAperture):
-    """
-    Represents an elliptical aperture that clips rays based on their position.
-
-    Attributes:
-        a (float): The semi-major axis of the ellipse.
-        b (float): The semi-minor axis of the ellipse.
-        offset_x (float): The x-coordinate of the aperture's center.
-        offset_y (float): The y-coordinate of the aperture's center.
-    """
-
-    def __init__(self, a, b, offset_x=0, offset_y=0):
-        super().__init__()
-        self.a = a
-        self.b = b
-        self.offset_x = offset_x
-        self.offset_y = offset_y
-
-    def clip(self, rays):
-        """
-        Clips the given rays based on their position.
-
-        Args:
-            rays (Rays): The rays to be clipped.
-        """
-        x = rays.x - self.offset_x
-        y = rays.y - self.offset_y
-        condition = (x**2 / self.a**2 + y**2 / self.b**2) > 1
-        rays.clip(condition)
-
-    def scale(self, scale_factor):
-        """
-        Scales the aperture by the given factor.
-
-        Args:
-            scale_factor (float): The factor by which to scale the aperture.
-        """
-        self.a *= scale_factor
-        self.b *= scale_factor
-        self.offset_x *= scale_factor
-        self.offset_y *= scale_factor
-
-    def to_dict(self):
-        """
-        Convert the aperture to a dictionary.
-
-        Returns:
-            dict: The dictionary representation of the aperture.
-        """
-        aperture_dict = super().to_dict()
-        aperture_dict['a'] = self.a
-        aperture_dict['b'] = self.b
-        aperture_dict['offset_x'] = self.offset_x
-        aperture_dict['offset_y'] = self.offset_y
-        return aperture_dict
-
-    @classmethod
-    def from_dict(cls, data):
-        """
-        Create an aperture from a dictionary representation.
-
-        Args:
-            data (dict): The dictionary representation of the aperture.
-
-        Returns:
-            EllipticalAperture: The aperture object.
-        """
-        return cls(data['a'], data['b'],
-                   data['offset_x'], data['offset_y'])
-
-
-class PolygonAperture(BaseAperture):
-    """
-    Represents a polygonal aperture that clips rays based on their position.
-
-    Attributes:
-        x (list or np.ndarray): x-coordinates of the polygon's vertices.
-        y (list or np.ndarray): y-coordinates of the polygon's vertices.
-        vertices (np.ndarray): Array-like of shape (n, 2) defining the
-            polygon vertices.
-    """
-
-    def __init__(self, x, y):
-        super().__init__()
-        self.x = x
-        self.y = y
-        self.vertices = np.column_stack((x, y))
-        self._path = Path(self.vertices)
-
-    def clip(self, rays):
-        """
-        Clips the given rays based on their position.
-
-        Args:
-            rays (Rays): The rays to be clipped.
-        """
-        points = np.column_stack((rays.x, rays.y))
-        inside = self._path.contains_points(points)
-        rays.clip(~inside)
-
-    def scale(self, scale_factor):
-        """
-        Scales the aperture by the given factor.
-
-        Args:
-            scale_factor (float): The factor by which to scale the aperture.
-        """
-        self.vertices *= scale_factor
-        self._path = Path(self.vertices)
-
-    def to_dict(self):
-        """
-        Convert the aperture to a dictionary.
-
-        Returns:
-            dict: The dictionary representation of the aperture.
-        """
-        aperture_dict = super().to_dict()
-        aperture_dict['x'] = self.x
-        aperture_dict['y'] = self.y
-        return aperture_dict
-
-    @classmethod
-    def from_dict(cls, data):
-        """
-        Create an aperture from a dictionary representation.
-
-        Args:
-            data (dict): The dictionary representation of the aperture.
-
-        Returns:
-            PolygonAperture: The aperture object.
-        """
-        return cls(data['x'], data['y'])
